pyproject.toml
[project]
name = "meme-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "autopep8>=2.3.2",
    "faiss-cpu",
    "gradio>=6.13.0",
    "numpy",
    "pillow",
    "pinecone",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "torch",
    "torchvision",
    "tqdm>=4.67.3",
    "uuid",
]
